profiling

+++ volume-cartographer/apps/VC3D/volume_viewers/CChunkedVolumeViewer.hpp
@@ -253,7 +253,7 @@ signals:
     void pointSelected(uint64_t pointId);
     void pointClicked(uint64_t pointId);
     void overlaysUpdated();
-    void renderFrameCompleted(std::uint64_t serial, qint64 workerElapsedMs);
+    void renderFrameCompleted(std::uint64_t serial, double workerElapsedMs);
     void sendSegmentationRadiusWheel(int steps, QPointF scenePoint, cv::Vec3f worldPos);
 
 private:
